+++ src/main/java/com/en_circle/slt/plugin/actions/swank/UndefineFunctionAction.java
@@ -2,12 +2,10 @@ package com.en_circle.slt.plugin.actions.swank;
 
 import com.en_circle.slt.plugin.SltBundle;
 import com.en_circle.slt.plugin.lisp.LispParserUtil;
-import com.en_circle.slt.plugin.lisp.lisp.LispElement;
-import com.en_circle.slt.plugin.lisp.lisp.LispString;
 import com.en_circle.slt.plugin.lisp.psi.LispSymbol;
 import com.en_circle.slt.plugin.services.lisp.LispEnvironmentService;
 import com.en_circle.slt.plugin.services.lisp.LispEnvironmentService.LispEnvironmentState;
-import com.en_circle.slt.plugin.swank.requests.Disassemble;
+import com.en_circle.slt.plugin.swank.requests.UndefineFunction;
 import com.intellij.codeInsight.hint.HintManager;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
@@ -15,8 +13,6 @@ import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.ex.EditorEx;
 import com.intellij.openapi.ui.Messages;
-import com.intellij.openapi.ui.popup.JBPopup;
-import com.intellij.openapi.ui.popup.JBPopupFactory;
 import com.intellij.psi.PsiDocumentManager;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
@@ -25,13 +21,10 @@ import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.swing.*;
-import java.awt.*;
 import java.util.Objects;
-import java.util.regex.Pattern;
 
-public class DisassembleAction extends SltSwankAction {
-    private static final Logger log = LoggerFactory.getLogger(DisassembleAction.class);
+public class UndefineFunctionAction extends SltSwankAction {
+    private static final Logger log = LoggerFactory.getLogger(UndefineFunctionAction.class);
 
     @Override
     public void actionPerformed(@NotNull AnActionEvent event) {
@@ -51,13 +44,12 @@ public class DisassembleAction extends SltSwankAction {
 
             if (element == null) {
                 ApplicationManager.getApplication().invokeLater(() -> HintManager.getInstance()
-                        .showErrorHint(editor, SltBundle.message("slt.ui.swank.disassemble.notsymbol")));
+                        .showErrorHint(editor, SltBundle.message("slt.ui.swank.undefinefunction.notsymbol")));
                 return;
             }
 
             String packageName = LispParserUtil.getPackage(element);
             String symbolname = element.getText();
-            String originalSymbol = symbolname;
             if (!symbolname.contains(":")) {
                 // assume it's in current package
                 symbolname = packageName + "::" + symbolname;
@@ -65,43 +57,12 @@ public class DisassembleAction extends SltSwankAction {
 
             try {
                 LispEnvironmentService.getInstance(editor.getProject())
-                        .sendToLisp(Disassemble.disassemble(symbolname, packageName,
-                                result -> SwingUtilities.invokeLater(() -> parseResult(result, editor))), false);
-                LispEnvironmentService.getInstance(editor.getProject())
-                        .refreshSymbolFromServer(packageName, originalSymbol);
+                        .sendToLisp(UndefineFunction.undefineFunction(symbolname, packageName), false);
             } catch (Exception e) {
                 log.warn(SltBundle.message("slt.error.start"), e);
                 Messages.showErrorDialog(editor.getProject(), e.getMessage(), SltBundle.message("slt.ui.errors.lisp.start"));
             }
         }
-    }
-
-    private void parseResult(LispElement result, EditorEx editor) {
-        String text = ((LispString) result).getValue();
-
-        JEditorPane editorPane = new JEditorPane("text/plain", text);
-        FontMetrics metrics = editorPane.getFontMetrics(editorPane.getFont());
-
-        String[] lines = text.split(Pattern.quote("\n"));
-        int max = 0;
-        for (String line : lines) {
-            max = Math.max(metrics.charsWidth(line.toCharArray(), 0, line.length()), max);
-        }
-
-        JPanel textPanel = new JPanel(new BorderLayout());
-        textPanel.add(editorPane, BorderLayout.CENTER);
-
-        JBPopup popup = JBPopupFactory.getInstance()
-                .createComponentPopupBuilder(textPanel, null)
-                .setProject(editor.getProject())
-                .setTitle(SltBundle.message("slt.ui.swank.disassemble.title"))
-                .setShowBorder(true)
-                .setMovable(true)
-                .setFocusable(true)
-                .setRequestFocus(true)
-                .setMinSize(new Dimension(max + 20, 0))
-                .createPopup();
-        popup.showInBestPositionFor(editor.getDataContext());
     }
 
     private boolean isSymbol(PsiElement element) {
@@ -113,5 +74,4 @@ public class DisassembleAction extends SltSwankAction {
         return LispEnvironmentService.getInstance(file.getProject())
                 .getState() == LispEnvironmentState.READY;
     }
-
 }