+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Refactor - extract lambda to function
 - Swank action - disassemble
+- Swank action - undefine function
 
 ### Changes